+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Auto Intensified` OP ability (requires TabletopTweaks-Base).
 - Added `Auto Elemental` OP abilities (requires TabletopTweaks-Base).
 - Added UMM option to allow for multiple Mythic OP abilities and Mythic special powers.
+- Added UMM option to restrict Exceptional Feats to the Isekai Protagonist Class.
 - Updated mod support for ExpandedContent.
 ### Changed
 - Reworked UMM settings.